Address PR review comments: server/client boundary + naming fixes
- Extract redirect logic from page.tsx into RootRedirect client component
- Show spinner through AuthGuard redirect instead of flashing blank
- Remove unnecessary 'use client' from workspace-card.tsx
- Rename WorkspaceTile to WorkspaceCard to match file name
